+++ src/main/kotlin/io/fdeitylink/kero/file/WritePxPack.kt
@@ -28,7 +28,7 @@ internal fun PxPack.toBytes() =
  * This does not comprise a fully valid PxPack file, just part of it.
  */
 private fun Head.toBytes(): ByteArray {
-    fun LayerProperties.toBytes() = tileset.toBytes() + visibilityType.type + scrollType.byte
+    fun LayerProperties.toBytes() = tileset.toBytes() + visibilityType.type.toByte() + scrollType.byte
 
     return Head.HEADER_STRING.toByteArray() +
